Convert jint to jlong for native pointers in GLFunctionTable to support 64 bit.
Corresponding Chromium change is: https://codereview.chromium.org/210023002

+++ chromium/java/com/android/webview/chromium/GraphicsUtils.java
@@ -17,15 +17,15 @@
 package com.android.webview.chromium;
 
 abstract class GraphicsUtils {
-    public static int getDrawSWFunctionTable() {
+    public static long getDrawSWFunctionTable() {
         return nativeGetDrawSWFunctionTable();
     }
 
-    public static int getDrawGLFunctionTable() {
+    public static long getDrawGLFunctionTable() {
         return nativeGetDrawGLFunctionTable();
     }
 
-    private static native int nativeGetDrawSWFunctionTable();
-    private static native int nativeGetDrawGLFunctionTable();
+    private static native long nativeGetDrawSWFunctionTable();
+    private static native long nativeGetDrawGLFunctionTable();
 
 }
